final mvc

+++ caelum/caelum_java_web_fj21/agenda/src/br/com/futrica/agenda/dao/ContatoDao.java
@@ -115,7 +115,7 @@ public class ContatoDao {
 		
 	}
 	
-	public Contato pesquisar(int id) {
+	public Contato pesquisar(long id) {
 		try{
 			Contato contato = new Contato();
 			PreparedStatement stmt = this.connection.prepareStatement("select * from contatos where id = " +id);
